+++ contao/languages/de/default.php
@@ -1,3 +1,6 @@
 <?php
 
 $GLOBALS['TL_LANG']['ERR']['turnstile'] = 'Die Sicherheitsüberprüfung (Cloudflare Turnstile) ist fehlgeschlagen. Bitte versuchen Sie es erneut.';
+
+// <noscript>-Hinweis im Widget. Auf '' setzen (Projekt-Override), um ihn auszublenden.
+$GLOBALS['TL_LANG']['MSC']['turnstileNoscript'] = 'Für die Sicherheitsprüfung muss JavaScript aktiviert sein.';